backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
llama-cpp-python==0.2.25
ctransformers==0.2.27
urllib3==1.26.16
numpy==1.24.3
markdown==3.5.1
bleach==6.1.0
Pygments==2.16.1
torch==2.0.1
scipy==1.10.1
sympy==1.12
boto3==1.28.76 #This is done for AWS Communication in the backend
psutil==5.9.8